Move and rename crates
https://matklad.github.io/2021/08/22/large-rust-workspaces.html

In general, having crate names that match the actual crate name is a lot
easier to understand. We had two different directories named "server,"
now they're easier to understand. As well as "common."

This also makes "wollongong" easier to remember, as now it's in the
actual name of the directories instead of being named something else.